+++ src/main/java/gov/usgs/volcanoes/swarm/picker/Event.java
@@ -1,0 +1,13 @@
+package gov.usgs.volcanoes.swarm.picker;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Event {
+  private List<Phase> phases;
+  
+  public Event() {
+    phases = new ArrayList<Phase>();
+  }
+
+}
