+++ src/main/java/gov/usgs/volcanoes/swarm/wave/WaveClipboardFrame.java
@@ -85,8 +85,8 @@ public class WaveClipboardFrame extends SwarmFrame {
 
   private JScrollPane scrollPane;
   private Box waveBox;
-  private final List<WaveViewPanel> waves;
-  private final Set<WaveViewPanel> selectedSet;
+  private final List<AbstractWavePanel> waves;
+  private final Set<AbstractWavePanel> selectedSet;
   private JToolBar toolbar;
   private JPanel mainPanel;
   private JLabel statusLabel;
@@ -117,7 +117,7 @@ public class WaveClipboardFrame extends SwarmFrame {
 
   private JPopupMenu popup;
 
-  private final Map<WaveViewPanel, Stack<double[]>> histories;
+  private final Map<AbstractWavePanel, Stack<double[]>> histories;
 
   private final HelicorderViewPanelListener linkListener;
 
@@ -132,11 +132,11 @@ public class WaveClipboardFrame extends SwarmFrame {
   private WaveClipboardFrame() {
     super("Wave Clipboard", true, true, true, false);
     this.setFocusable(true);
-    selectedSet = new HashSet<WaveViewPanel>();
+    selectedSet = new HashSet<AbstractWavePanel>();
     saveAllDateFormat = new SimpleDateFormat("yyyyMMddHHmmss");
     saveAllDateFormat.setTimeZone(TimeZone.getTimeZone("GMT"));
-    waves = new ArrayList<WaveViewPanel>();
-    histories = new HashMap<WaveViewPanel, Stack<double[]>>();
+    waves = new ArrayList<AbstractWavePanel>();
+    histories = new HashMap<AbstractWavePanel, Stack<double[]>>();
     createUI();
     linkListener = new HelicorderViewPanelListener() {
       public void insetCreated(final double st, final double et) {
@@ -286,12 +286,12 @@ public class WaveClipboardFrame extends SwarmFrame {
 
       int height = 0;
       final int width = waves.get(0).getWidth();
-      for (final WaveViewPanel panel : waves)
+      for (final AbstractWavePanel panel : waves)
         height += panel.getHeight();
 
       final BufferedImage image = new BufferedImage(width, height, BufferedImage.TYPE_4BYTE_ABGR);
       final Graphics g = image.getGraphics();
-      for (final WaveViewPanel panel : waves) {
+      for (final AbstractWavePanel panel : waves) {
         panel.paint(g);
         g.translate(0, panel.getHeight());
       }
@@ -428,7 +428,7 @@ public class WaveClipboardFrame extends SwarmFrame {
       private static final long serialVersionUID = 1L;
 
       public void actionPerformed(final ActionEvent e) {
-        for (final WaveViewPanel wave : waves)
+        for (final AbstractWavePanel wave : waves)
           select(wave);
       }
     });
@@ -462,7 +462,7 @@ public class WaveClipboardFrame extends SwarmFrame {
 
     WaveViewTime.addTimeListener(new TimeListener() {
       public void timeChanged(final double j2k) {
-        for (final WaveViewPanel panel : waves) {
+        for (final AbstractWavePanel panel : waves) {
           if (panel != null)
             panel.setCursorMark(j2k);
         }
@@ -471,7 +471,7 @@ public class WaveClipboardFrame extends SwarmFrame {
 
     selectListener = new WaveViewPanelAdapter() {
       @Override
-      public void mousePressed(final WaveViewPanel src, final MouseEvent e,
+      public void mousePressed(final AbstractWavePanel src, final MouseEvent e,
           final boolean dragging) {
         requestFocusInWindow();
         final int thisIndex = getWaveIndex(src);
@@ -500,11 +500,11 @@ public class WaveClipboardFrame extends SwarmFrame {
       }
 
       @Override
-      public void waveZoomed(final WaveViewPanel src, final double st, final double et,
+      public void waveZoomed(final AbstractWavePanel src, final double st, final double et,
           final double nst, final double net) {
         final double[] t = new double[] {st, et};
         addHistory(src, t);
-        for (final WaveViewPanel wvp : selectedSet) {
+        for (final AbstractWavePanel wvp : selectedSet) {
           if (wvp != src) {
             addHistory(wvp, t);
             wvp.zoom(nst, net);
@@ -513,7 +513,7 @@ public class WaveClipboardFrame extends SwarmFrame {
       }
 
       @Override
-      public void waveClosed(final WaveViewPanel src) {
+      public void waveClosed(final AbstractWavePanel src) {
         remove(src);
       }
     };
@@ -598,7 +598,7 @@ public class WaveClipboardFrame extends SwarmFrame {
 
   private class SaveActionListener implements ActionListener {
     public void actionPerformed(final ActionEvent e) {
-      final WaveViewPanel selected = getSingleSelected();
+      final AbstractWavePanel selected = getSingleSelected();
       if (selected == null)
         return;
 
@@ -708,7 +708,7 @@ public class WaveClipboardFrame extends SwarmFrame {
             return;
           if (!f.exists())
             f.mkdir();
-          for (final WaveViewPanel wvp : waves) {
+          for (final AbstractWavePanel wvp : waves) {
             Wave sw = wvp.getWave();
 
             if (sw != null) {
@@ -811,20 +811,20 @@ public class WaveClipboardFrame extends SwarmFrame {
   }
 
   public synchronized void sortChannelsByNearest() {
-    final WaveViewPanel p = getSingleSelected();
+    final AbstractWavePanel p = getSingleSelected();
     if (p == null)
       return;
 
-    final ArrayList<WaveViewPanel> sorted = new ArrayList<WaveViewPanel>(waves.size());
-    for (final WaveViewPanel wave : waves)
+    final ArrayList<AbstractWavePanel> sorted = new ArrayList<AbstractWavePanel>(waves.size());
+    for (final AbstractWavePanel wave : waves)
       sorted.add(wave);
 
     final Metadata smd = swarmConfig.getMetadata(p.getChannel());
     if (smd == null || Double.isNaN(smd.getLongitude()) || Double.isNaN(smd.getLatitude()))
       return;
 
-    Collections.sort(sorted, new Comparator<WaveViewPanel>() {
-      public int compare(final WaveViewPanel wvp1, final WaveViewPanel wvp2) {
+    Collections.sort(sorted, new Comparator<AbstractWavePanel>() {
+      public int compare(final AbstractWavePanel wvp1, final AbstractWavePanel wvp2) {
         Metadata md = swarmConfig.getMetadata(wvp1.getChannel());
         final double d1 = smd.distanceTo(md);
         md = swarmConfig.getMetadata(wvp2.getChannel());
@@ -834,24 +834,24 @@ public class WaveClipboardFrame extends SwarmFrame {
     });
 
     removeWaves();
-    for (final WaveViewPanel wave : sorted)
+    for (final AbstractWavePanel wave : sorted)
       addWave(wave);
     select(p);
   }
 
-  public synchronized WaveViewPanel getSingleSelected() {
+  public synchronized AbstractWavePanel getSingleSelected() {
     if (selectedSet.size() != 1)
       return null;
 
-    WaveViewPanel p = null;
-    for (final WaveViewPanel panel : selectedSet)
+    AbstractWavePanel p = null;
+    for (final AbstractWavePanel panel : selectedSet)
       p = panel;
 
     return p;
   }
 
   public synchronized void syncChannels() {
-    final WaveViewPanel p = getSingleSelected();
+    final AbstractWavePanel p = getSingleSelected();
     if (p == null)
       return;
 
@@ -862,11 +862,11 @@ public class WaveClipboardFrame extends SwarmFrame {
     final SwingWorker worker = new SwingWorker() {
       @Override
       public Object construct() {
-        List<WaveViewPanel> copy = null;
+        List<AbstractWavePanel> copy = null;
         synchronized (WaveClipboardFrame.this) {
-          copy = new ArrayList<WaveViewPanel>(waves);
+          copy = new ArrayList<AbstractWavePanel>(waves);
         }
-        for (final WaveViewPanel wvp : copy) {
+        for (final AbstractWavePanel wvp : copy) {
           if (wvp != p) {
             if (wvp.getDataSource() != null) {
               addHistory(wvp, new double[] {wvp.getStartTime(), wvp.getEndTime()});
@@ -899,7 +899,7 @@ public class WaveClipboardFrame extends SwarmFrame {
     // return selected;
   }
 
-  public synchronized void addWave(final WaveViewPanel p) {
+  public synchronized void addWave(final AbstractWavePanel p) {
     p.addListener(selectListener);
     p.setOffsets(54, 8, 21, 19);
     p.setAllowClose(true);
@@ -916,7 +916,7 @@ public class WaveClipboardFrame extends SwarmFrame {
     waveBox.validate();
   }
 
-  private synchronized void deselect(final WaveViewPanel p) {
+  private synchronized void deselect(final AbstractWavePanel p) {
     selectedSet.remove(p);
     waveToolbar.removeSettings(p.getSettings());
     p.setBackgroundColor(BACKGROUND_COLOR);
@@ -930,7 +930,7 @@ public class WaveClipboardFrame extends SwarmFrame {
       deselect(p);
   }
 
-  private synchronized void select(final WaveViewPanel p) {
+  private synchronized void select(final AbstractWavePanel p) {
     if (p == null || selectedSet.contains(p))
       return;
 
@@ -963,7 +963,7 @@ public class WaveClipboardFrame extends SwarmFrame {
     repaint();
   }
 
-  protected int getWaveIndex(final WaveViewPanel p) {
+  protected int getWaveIndex(final AbstractWavePanel p) {
     int i = 0;
     for (i = 0; i < waveBox.getComponentCount(); i++) {
       if (p == waveBox.getComponent(i))
@@ -979,7 +979,7 @@ public class WaveClipboardFrame extends SwarmFrame {
   }
 
   public synchronized void moveDown() {
-    final WaveViewPanel p = getSingleSelected();
+    final AbstractWavePanel p = getSingleSelected();
     if (p == null)
       return;
 
@@ -996,7 +996,7 @@ public class WaveClipboardFrame extends SwarmFrame {
   }
 
   public synchronized void moveUp() {
-    final WaveViewPanel p = getSingleSelected();
+    final AbstractWavePanel p = getSingleSelected();
     if (p == null)
       return;
 
@@ -1017,7 +1017,7 @@ public class WaveClipboardFrame extends SwarmFrame {
       @Override
       public Object construct() {
         final int w = scrollPane.getViewport().getSize().width;
-        for (final WaveViewPanel wave : waves) {
+        for (final AbstractWavePanel wave : waves) {
           wave.setSize(w, calculateWaveHeight());
           wave.createImage();
         }
@@ -1044,7 +1044,7 @@ public class WaveClipboardFrame extends SwarmFrame {
     repaint();
   }
 
-  private void addHistory(final WaveViewPanel wvp, final double[] t) {
+  private void addHistory(final AbstractWavePanel wvp, final double[] t) {
     Stack<double[]> history = histories.get(wvp);
     if (history == null) {
       history = new Stack<double[]>();
@@ -1053,7 +1053,7 @@ public class WaveClipboardFrame extends SwarmFrame {
     history.push(t);
   }
 
-  public void gotoTime(final WaveViewPanel wvp, String t) {
+  public void gotoTime(final AbstractWavePanel wvp, String t) {
     double j2k = Double.NaN;
     try {
       if (t.length() == 12)
@@ -1086,11 +1086,11 @@ public class WaveClipboardFrame extends SwarmFrame {
   }
 
   public void gotoTime(final String t) {
-    for (final WaveViewPanel p : selectedSet)
+    for (final AbstractWavePanel p : selectedSet)
       gotoTime(p, t);
   }
 
-  public void scaleTime(final WaveViewPanel wvp, final double pct) {
+  public void scaleTime(final AbstractWavePanel wvp, final double pct) {
     final double st = wvp.getStartTime();
     final double et = wvp.getEndTime();
     final double[] t = new double[] {st, et};
@@ -1103,11 +1103,11 @@ public class WaveClipboardFrame extends SwarmFrame {
   }
 
   public void scaleTime(final double pct) {
-    for (final WaveViewPanel p : selectedSet)
+    for (final AbstractWavePanel p : selectedSet)
       scaleTime(p, pct);
   }
 
-  public void back(final WaveViewPanel wvp) {
+  public void back(final AbstractWavePanel wvp) {
     final Stack<double[]> history = histories.get(wvp);
     if (history == null || history.empty())
       return;
@@ -1117,11 +1117,11 @@ public class WaveClipboardFrame extends SwarmFrame {
   }
 
   public void back() {
-    for (final WaveViewPanel p : selectedSet)
+    for (final AbstractWavePanel p : selectedSet)
       back(p);
   }
 
-  private void shiftTime(final WaveViewPanel wvp, final double pct) {
+  private void shiftTime(final AbstractWavePanel wvp, final double pct) {
     final double st = wvp.getStartTime();
     final double et = wvp.getEndTime();
     final double[] t = new double[] {st, et};
@@ -1133,12 +1133,12 @@ public class WaveClipboardFrame extends SwarmFrame {
   }
 
   public void shiftTime(final double pct) {
-    for (final WaveViewPanel p : selectedSet)
+    for (final AbstractWavePanel p : selectedSet)
       shiftTime(p, pct);
   }
 
   public void repositionWaves(final double st, final double et) {
-    for (final WaveViewPanel wave : waves) {
+    for (final AbstractWavePanel wave : waves) {
       fetchNewWave(wave, st, et);
     }
   }
@@ -1148,7 +1148,7 @@ public class WaveClipboardFrame extends SwarmFrame {
   }
 
   // TODO: This isn't right, this should be a method of waveviewpanel
-  private void fetchNewWave(final WaveViewPanel wvp, final double nst, final double net) {
+  private void fetchNewWave(final AbstractWavePanel wvp, final double nst, final double net) {
     final SwingWorker worker = new SwingWorker() {
       @Override
       public Object construct() {
